resolved the pr comments
