+++ src/apps/frontend/pages/tasks/index.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
 import toast from 'react-hot-toast';
 
-import { TaskContainer } from '../../components';
 import { AsyncError } from '../../types';
+
+import TaskContainer from './task-container';
 
 const Tasks: React.FC = () => {
   const onError = (error: AsyncError) => {
@@ -10,9 +11,9 @@ const Tasks: React.FC = () => {
   };
 
   return (
-      <div className='h-screen overflow-y-auto p-5'>
-        <TaskContainer onError={onError} />
-      </div>
+    <div className="h-screen overflow-y-auto p-5">
+      <TaskContainer onError={onError} />
+    </div>
   );
 };
